requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.7.1
requests==2.32.3
python-dotenv==1.0.1
serpapi==0.1.5
pytest==8.2.1
